DSND-3113: Renamed psc package to psc_statements to avoid confusion
