--- validation/schema_testing/psc/pscStatementDeltaSchema.go
@@ -1,1 +1,0 @@
-package psc
