+++ validation/schema_testing/psc_statements/pscStatementDeltaSchema.go
@@ -1,0 +1,1 @@
+package psc_statements
