+++ validation/schema_testing/psc_statements/request_bodies/delete_ok_request_body
