+++ validation/schema_testing/psc_statements/response_bodies/delete_error_response_body
