+++ validation/schema_testing/psc_statements/response_bodies/required_error_response_body
