+++ validation/schema_testing/psc_statements/response_bodies/type_error_response_body
